Switch from using proxy to building each page
Using proxies to build from a YAML file might seem a good idea but it
has some problems. Trying to incorporate more involved content into the
YAML file is doable, but horrible and is likely to result in
maintenance difficulties. So, back to plain old a file for each page.