Update index.html.erb.md
Just minor wording edits.

Also changed the order of the activities. The goals/aims/objectives will need to be established before a business case can be developed. And the business case needs to be in place with the goals/aims/objectives included before you can raise awareness of what you are trying to accomplish and communicate them to gain organizational wide buy-in.