Update assess_the_situation.html.erb.md
Wording edits. (feel free to take them or leave them)

Also, I don't like this sentence "* Web content policies and processes, for example related to editing and publishing;" but I don't know what to replace it with. I like the first part and the example, but it just does not flow well in the bulleted list.