Update establish_monitoring_framework.html.erb.md
Removed lots of IMO unnecessary wordiness and repetition. Also tangential stuff. The very last "review early on" is excellent but not really here IMO. I moved it to "Review websites" though this could also go into implementation. This is more of a "how/when to evaluate" than establishing a framework.